Move WitnessRpcHub
https://github.com/Not-Endurance/not-timing-system/issues/433
